fix(agents): ParseAgent ExtractString numPages/confidence are uint8, not uint256
Grounded against docs.somnia.network/agents/base-agents/llm-parse-website: the live
parse-website agent's selector is ExtractString(...,bool,uint8,uint8). Our uint256 width
produced selector 0x15385dfb instead of the correct 0xc2dd1a7a — every real parse request
would have TimedOut (mock tests didn't catch it; the mock ignores the selector). Fixed in
IParseAgent, SomniaAI.encodeExtractString, ParseAgent.ExtractParams, and the SDK parseAgentAbi.

Also confirmed against official docs (verbatim match, no change needed): IAgentRequester,
IAgentRequesterHandler, Response/Request structs, ResponseStatus enum, deposit formula,
ILlmAgent.inferString/inferNumber. Contracts 131 tests pass; SDK 37 pass.

+++ packages/contracts/src/agents/ParseAgent.sol
@@ -30,8 +30,8 @@ contract ParseAgent is AgentCompute {
     /// @param prompt      extraction instruction
     /// @param url         page to read
     /// @param resolveUrl  true = domain-search mode; false = direct scrape (numPages → 1)
-    /// @param numPages    pages to read
-    /// @param confidenceThreshold 0–100 gate below which extraction fails
+    /// @param numPages    pages to read (uint8 — matches the agent's ExtractString selector)
+    /// @param confidenceThreshold 0–100 gate below which extraction fails (uint8)
     struct ExtractParams {
         string key;
         string description;
@@ -39,8 +39,8 @@ contract ParseAgent is AgentCompute {
         string prompt;
         string url;
         bool resolveUrl;
-        uint256 numPages;
-        uint256 confidenceThreshold;
+        uint8 numPages;
+        uint8 confidenceThreshold;
     }
 
     event ExtractionRequested(uint256 indexed requestId, string url, string key);
